app/requirements.txt
# ===========================================
# EcoSort-AI Dependencies
# ===========================================

# Core Framework
streamlit>=1.40.0

# AI/ML
google-generativeai>=0.8.0

# Image Processing
Pillow>=10.0.0

# Azure Integration
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0
httpx>=0.27.0

# ===========================================
# Development Dependencies (install separately)
# ===========================================
# pytest>=8.0.0
# pytest-cov>=4.1.0
# pytest-asyncio>=0.23.0
# black>=24.0.0
# ruff>=0.4.0
# mypy>=1.8.0
# bandit>=1.7.0
# safety>=3.0.0
